requirements.txt
# Backend requirements
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
pydantic-settings
tavily-python
python-multipart
pymilvus
sentence-transformers
torch
# uvloop and httptools are optional and not supported on Windows
# uvloop
# httptools
